pyproject.toml
[project]
name = "plsconvert"
version = "0.1.5"
description = "Convert any file to any other."
readme = "README.md"
authors = [{ name = "ThorTuwy", email = "tuwy@tuwy.win" }, { name = "Víctor", email = "victordlp8@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "halo>=0.0.31",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
docs = ["pdf2docx>=0.5.8", "pandas>=2.3.1", "openpyxl>=3.1.5"]

audio = ["matplotlib>=3.10.3", "scipy>=1.16.0", "pyttsx3>=2.99", "midi2audio>=0.1.1"]

3d = ["trimesh>=4.5.1", "moderngl>=5.12.0", "pyrr>=0.10.3", "pillow>=11.3.0", "imageio>=2.36.1", "imageio-ffmpeg>=0.5.1", "pygltflib>=1.16.3"]

graph = ["matplotlib>=3.10.3", "networkx>=3.0", "netgraph>=4.13.2"]

config = ["pyyaml>=6.0.2", "tomlkit>=0.13.3"]

ai = ["rapidocr>=3.2.0", "onnxruntime>=1.22.1", "pillow>=11.3.0"]

noAI = ["plsconvert[docs]", "plsconvert[audio]", "plsconvert[config]"]

all = ["plsconvert[docs]", "plsconvert[audio]", "plsconvert[config]", "plsconvert[ai]", "plsconvert[3d]"]

dev = ["plsconvert[all]", "plsconvert[graph]"]

[project.urls]
Homepage = "https://github.com/ThorTuwy/plsconvert"
"Bug Tracker" = "https://github.com/ThorTuwy/plsconvert/issues"

[project.scripts]
plsconvert = "plsconvert:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
